requirements.txt
fastapi
uvicorn[standard]
httpx
google-genai>=1.0.0
python-dotenv